7-4-2022
This is the entering of all the data for this dataset, and I'm honestly slightly intimidated by it. There's a lot here. I'm still feeling out what I want to do with it, but for now, here is just the data. I'll have the notebook uploaded sometime this week.
There is also a set of images from all of these brands that I could use, but I'm hesitating. I have to look into what I can do with it, first.